@@ -1,10 +1,10 @@
-import 'package:app/biodata/appbar.dart';
+import 'package:frontend_flutter/biodata/appbar.dart';
 import 'package:flutter/material.dart';
 
 class SubIntroWidget extends StatelessWidget {
   final VoidCallback onBack;
 
-  SubIntroWidget({required this.onBack});
+  const SubIntroWidget({super.key, required this.onBack});
 
   @override
   Widget build(BuildContext context) {
@@ -12,7 +12,7 @@ class SubIntroWidget extends StatelessWidget {
       child: Column(
         mainAxisAlignment: MainAxisAlignment.start,
         children: [
-          Appbar(onBack: onBack,),
+          Appbar(onBack: onBack),
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
@@ -23,29 +23,67 @@ class SubIntroWidget extends StatelessWidget {
                   children: [
                     Text(
                       "Let us know you",
-                      style: TextStyle(fontSize: 36, fontWeight: FontWeight.bold),
+                      style: TextStyle(
+                        fontSize: 36,
+                        fontWeight: FontWeight.bold,
+                        fontFamily: 'Nunito',
+                      ),
                     ),
                     SizedBox(height: 5),
                     Text(
                       "Introduce yourself by letting us know who you're so that we can provide the best service.",
-                      style: TextStyle(fontSize: 16, color: Colors.grey),
+                      style: TextStyle(
+                        fontSize: 16,
+                        color: Colors.grey,
+                        fontFamily: 'Nunito',
+                      ),
                     ),
                     SizedBox(height: 20),
-                    Text("Name", style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold)),
+                    Text(
+                      "Name",
+                      style: TextStyle(
+                        fontSize: 18,
+                        fontWeight: FontWeight.bold,
+                        fontFamily: 'Nunito',
+                      ),
+                    ),
                     SizedBox(height: 8),
                     Row(
                       children: [
-                        Expanded(child: _buildTextField("First Name", Icons.person_outline)),
+                        Expanded(
+                          child: _buildTextField(
+                            "First Name",
+                            Icons.person_outline,
+                          ),
+                        ),
                         SizedBox(width: 10),
-                        Expanded(child: _buildTextField("Last Name", Icons.person_outline)),
+                        Expanded(
+                          child: _buildTextField(
+                            "Last Name",
+                            Icons.person_outline,
+                          ),
+                        ),
                       ],
                     ),
                     SizedBox(height: 20),
-                    Text("Address", style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold)),
+                    Text(
+                      "Address",
+                      style: TextStyle(
+                        fontSize: 18,
+                        fontWeight: FontWeight.bold,
+                        fontFamily: 'Nunito',
+                      ),
+                    ),
                     SizedBox(height: 8),
-                    _buildTextField("Number, Street Name", Icons.location_on_outlined),
+                    _buildTextField(
+                      "Number, Street Name",
+                      Icons.location_on_outlined,
+                    ),
                     SizedBox(height: 10),
-                    _buildTextField("Flat Number/Building Number", Icons.apartment),
+                    _buildTextField(
+                      "Flat Number/Building Number",
+                      Icons.apartment,
+                    ),
                     SizedBox(height: 10),
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -57,9 +95,19 @@ class SubIntroWidget extends StatelessWidget {
                     ),
                     SizedBox(height: 20),
                     // Meal Drop Location
-                    Text("Where can we leave your meals?", style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold)),
+                    Text(
+                      "Where can we leave your meals?",
+                      style: TextStyle(
+                        fontSize: 16,
+                        fontWeight: FontWeight.bold,
+                        fontFamily: 'Nunito',
+                      ),
+                    ),
                     SizedBox(height: 8),
-                    _buildTextField("Place to leave order", Icons.location_on_outlined),
+                    _buildTextField(
+                      "Place to leave order",
+                      Icons.location_on_outlined,
+                    ),
                     SizedBox(height: 20),
                   ],
                 ),
@@ -77,9 +125,14 @@ class SubIntroWidget extends StatelessWidget {
                       backgroundColor: Color(0xFFCF353F),
                       foregroundColor: Colors.white,
                       padding: EdgeInsets.symmetric(vertical: 14),
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(8),
+                      ),
                     ),
-                    child: Text("Continue", style: TextStyle(fontSize: 16)),
+                    child: Text(
+                      "Continue",
+                      style: TextStyle(fontSize: 16, fontFamily: 'Nunito'),
+                    ),
                   ),
                 ),
               ),
@@ -99,7 +152,6 @@ class SubIntroWidget extends StatelessWidget {
         filled: true,
         fillColor: Colors.white,
         contentPadding: EdgeInsets.symmetric(vertical: 14),
-
       ),
     );
   }
@@ -114,8 +166,8 @@ class SubIntroWidget extends StatelessWidget {
       child: Row(
         children: [
           Radio(value: label, groupValue: null, onChanged: (value) {}),
-          Text(label, style: TextStyle(fontSize: 18),),
-          SizedBox(width: 15,)
+          Text(label, style: TextStyle(fontSize: 18, fontFamily: 'Nunito')),
+          SizedBox(width: 15),
         ],
       ),
     );
