@@ -1,12 +1,13 @@
-import 'package:app/biodata/otp.dart';
-import 'package:app/biodata/phone.dart';
-import 'package:app/biodata/postcode.dart';
-import 'package:app/biodata/restro_intro.dart';
-import 'package:app/biodata/role_select.dart';
-import 'package:app/biodata/sub_intro.dart';
+import 'package:frontend_flutter/biodata/phone.dart';
+import 'package:frontend_flutter/biodata/postcode.dart';
+import 'package:frontend_flutter/biodata/restro_intro.dart';
+import 'package:frontend_flutter/biodata/role_select.dart';
+import 'package:frontend_flutter/biodata/sub_intro.dart';
 import 'package:flutter/material.dart';
 
 class BioDataPage extends StatefulWidget {
+  const BioDataPage({super.key});
+
   @override
   _BioDataPageState createState() => _BioDataPageState();
 }
@@ -38,7 +39,7 @@ class _BioDataPageState extends State<BioDataPage> {
           physics: NeverScrollableScrollPhysics(),
           children: [
             RoleSelectionWidget(
-              onContinue: (selectedRole){
+              onContinue: (selectedRole) {
                 print("$selectedRole");
                 setState(() {
                   _selectedRole = selectedRole;
@@ -46,19 +47,12 @@ class _BioDataPageState extends State<BioDataPage> {
                 _nextPage();
               },
             ),
-            PhoneNumberEntryWidget(
-              onBack: _prevPage,
-              onContinue: _nextPage,
-            ),
-            OTPWidget(
-              onBack: _prevPage,
-              onContinue: _nextPage,
-            ),
-            PostcodeWidget(
-              onBack: _prevPage,
-              onContinue: _nextPage
-            ),
-            if (_selectedRole == 0) SubIntroWidget(onBack: _prevPage,) else RestroIntroWidget(onBack: _prevPage),
+            PhoneNumberEntryWidget(onBack: _prevPage, onContinue: _nextPage),
+            PostcodeWidget(onBack: _prevPage, onContinue: _nextPage),
+            if (_selectedRole == 0)
+              SubIntroWidget(onBack: _prevPage)
+            else
+              RestroIntroWidget(onBack: _prevPage),
           ],
         ),
       ),
