@@ -1,8 +1,10 @@
-import 'package:app/custom_button.dart';
+import 'package:frontend_flutter/custom_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_phosphor_icons/flutter_phosphor_icons.dart';
 
 class AddMealsPage extends StatefulWidget {
+  const AddMealsPage({super.key});
+
   @override
   _AddMealsPageState createState() => _AddMealsPageState();
 }
@@ -10,15 +12,16 @@ class AddMealsPage extends StatefulWidget {
 class _AddMealsPageState extends State<AddMealsPage> {
   int mealCount = 0;
   String selectedDay = "Sun";
-  int selectedIndex = 0 ;
+  int selectedIndex = 0;
   List<Map<String, dynamic>> meals = [
     {
       "image": "assets/peri_peri.png",
       "tags": ["Dairy Free", "Gluten Free", "Halal"],
-      "name": "Peri Peri Chicken Breast, Peri Pilaf Rice, Mediterran Veg and Homemade Peri Sauce",
+      "name":
+          "Peri Peri Chicken Breast, Peri Pilaf Rice, Mediterran Veg and Homemade Peri Sauce",
       "details": "423 kcal | 40.9g protein",
       "price": 49.86,
-      "count": 0
+      "count": 0,
     },
     {
       "image": "assets/bbq.png",
@@ -26,8 +29,8 @@ class _AddMealsPageState extends State<AddMealsPage> {
       "name": "Vegan Protein Bowl",
       "details": "390 kcal | 35.2g protein",
       "price": 39.99,
-      "count": 0
-    }
+      "count": 0,
+    },
   ];
 
   @override
@@ -37,20 +40,27 @@ class _AddMealsPageState extends State<AddMealsPage> {
       appBar: AppBar(
         title: Text(
           "Add Meals",
-          style: TextStyle(fontSize: 18, fontWeight: FontWeight.w500),
+          style: TextStyle(
+            fontSize: 18,
+            fontWeight: FontWeight.w500,
+            fontFamily: 'Nunito',
+          ),
         ),
         centerTitle: true,
         backgroundColor: Colors.white,
         elevation: 0,
         leading: IconButton(
-          icon: Icon(PhosphorIcons.caret_left, size: 24, ),
+          icon: Icon(PhosphorIcons.caret_left, size: 24),
           onPressed: () => Navigator.pop(context),
+        ),
+        bottom: PreferredSize(
+          preferredSize: Size.fromHeight(64),
+          child: _buildDaySelector(),
         ),
       ),
       body: Column(
         children: [
-          _buildDaySelector(),
-          SizedBox(height: 16,),
+          SizedBox(height: 16),
           _buildMealTypeTabs(),
           SizedBox(
             height: 518,
@@ -58,9 +68,9 @@ class _AddMealsPageState extends State<AddMealsPage> {
               padding: const EdgeInsets.symmetric(horizontal: 16),
               child: Column(
                 children: [
-                  SizedBox(height: 8,),
+                  SizedBox(height: 8),
                   _buildSearchBar(),
-                  SizedBox(height: 8,),
+                  SizedBox(height: 8),
                   Expanded(child: _buildMealList()),
                 ],
               ),
@@ -76,38 +86,46 @@ class _AddMealsPageState extends State<AddMealsPage> {
     List<String> days = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
     return Container(
       padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-      color: Colors.white,
+      //color: Colors.white,
       width: double.infinity,
       child: SingleChildScrollView(
         scrollDirection: Axis.horizontal,
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: days.map((day) {
-            return Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 8.0),
-              child: GestureDetector(
-                onTap: () {
-                  setState(() {
-                    selectedDay = day;
-                  });
-                },
-                child: Container(
-                  padding: EdgeInsets.all(8),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(4),
-                    color: selectedDay == day ? Color(0xFFCF353F) : Color(0xFFF9F9F9),
-                  ),
-                  child: Text(
-                    day,
-                    style: TextStyle(
-                      fontSize: 18,
-                      color: selectedDay == day ? Colors.white : Color(0xFF2B2D42),
+          children:
+              days.map((day) {
+                return Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                  child: GestureDetector(
+                    onTap: () {
+                      setState(() {
+                        selectedDay = day;
+                      });
+                    },
+                    child: Container(
+                      padding: EdgeInsets.all(8),
+                      decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(4),
+                        color:
+                            selectedDay == day
+                                ? Color(0xFFCF353F)
+                                : Color(0xFFF9F9F9),
+                      ),
+                      child: Text(
+                        day,
+                        style: TextStyle(
+                          fontSize: 18,
+                          color:
+                              selectedDay == day
+                                  ? Colors.white
+                                  : Color(0xFF2B2D42),
+                          fontFamily: 'Nunito',
+                        ),
+                      ),
                     ),
                   ),
-                ),
-              ),
-            );
-          }).toList(),
+                );
+              }).toList(),
         ),
       ),
     );
@@ -121,25 +139,50 @@ class _AddMealsPageState extends State<AddMealsPage> {
         height: 48,
         width: double.infinity,
         decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(10)
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(10),
         ),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: [
             _buildTab(
-                Text("Lunch", style: TextStyle(color: selectedIndex == 0 ? Color(0xFFCF353F) : Color(0xFF49454F), fontSize: 14, fontWeight: selectedIndex == 0 ? FontWeight.w600 : FontWeight.normal),),
-                0
+              Text(
+                "Lunch",
+                style: TextStyle(
+                  color:
+                      selectedIndex == 0
+                          ? Color(0xFFCF353F)
+                          : Color(0xFF49454F),
+                  fontSize: 14,
+                  fontWeight:
+                      selectedIndex == 0 ? FontWeight.w600 : FontWeight.normal,
+                  fontFamily: 'Nunito',
+                ),
+              ),
+              0,
             ),
             _buildTab(
-                Text("Dinner", style: TextStyle(color: selectedIndex == 1 ? Color(0xFFCF353F) : Color(0xFF49454F), fontSize: 14, fontWeight: selectedIndex == 1 ? FontWeight.w600 : FontWeight.normal),),
-                1
+              Text(
+                "Dinner",
+                style: TextStyle(
+                  color:
+                      selectedIndex == 1
+                          ? Color(0xFFCF353F)
+                          : Color(0xFF49454F),
+                  fontSize: 14,
+                  fontWeight:
+                      selectedIndex == 1 ? FontWeight.w600 : FontWeight.normal,
+                  fontFamily: 'Nunito',
+                ),
+              ),
+              1,
             ),
           ],
         ),
       ),
     );
   }
+
   Widget _buildTab(Widget tabContent, int index) {
     return GestureDetector(
       onTap: () {
@@ -149,10 +192,7 @@ class _AddMealsPageState extends State<AddMealsPage> {
       },
       child: Column(
         children: [
-          SizedBox(
-            width: 180,
-            child: Center(child: tabContent),
-          ),
+          SizedBox(width: 180, child: Center(child: tabContent)),
           if (selectedIndex == index)
             Container(
               margin: const EdgeInsets.only(top: 10),
@@ -166,7 +206,7 @@ class _AddMealsPageState extends State<AddMealsPage> {
   }
 
   Widget _buildSearchBar() {
-    return  Container(
+    return Container(
       height: 48,
       decoration: BoxDecoration(
         color: Colors.white,
@@ -175,18 +215,18 @@ class _AddMealsPageState extends State<AddMealsPage> {
       ),
       child: Row(
         children: [
-          SizedBox(width: 8,),
+          SizedBox(width: 8),
           const Expanded(
             child: TextField(
               decoration: InputDecoration(
                 hintText: "Search across the menu",
                 border: InputBorder.none,
-                hintStyle: TextStyle(color: Colors.grey),
+                hintStyle: TextStyle(color: Colors.grey, fontFamily: 'Nunito'),
               ),
             ),
           ),
           IconButton(
-            onPressed: (){},
+            onPressed: () {},
             icon: Icon(PhosphorIcons.magnifying_glass, color: Colors.black),
           ),
           const SizedBox(width: 8),
@@ -200,18 +240,32 @@ class _AddMealsPageState extends State<AddMealsPage> {
       itemCount: meals.length,
       itemBuilder: (context, index) {
         final meal = meals[index];
-        List<Color> tagColors = [Colors.orange, Colors.blue, Colors.green, Colors.purple, Colors.red, Colors.teal];
-        List<Color> textColor = [Color(0xFF2B2D42), Colors.white, Colors.white, Colors.white, Colors.white, Colors.white];
+        List<Color> tagColors = [
+          Colors.orange,
+          Colors.blue,
+          Colors.green,
+          Colors.purple,
+          Colors.red,
+          Colors.teal,
+        ];
+        List<Color> textColor = [
+          Color(0xFF2B2D42),
+          Colors.white,
+          Colors.white,
+          Colors.white,
+          Colors.white,
+          Colors.white,
+        ];
 
         return GestureDetector(
-          onTap: (){
+          onTap: () {
             print("Clicked");
           },
           child: Container(
             margin: EdgeInsets.symmetric(vertical: 12),
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(12),
-              color: Colors.white
+              color: Colors.white,
             ),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
@@ -231,50 +285,94 @@ class _AddMealsPageState extends State<AddMealsPage> {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       Row(
-                        children: meal["tags"].asMap().entries.map<Widget>((entry) {
-                          int tagIndex = entry.key;
-                          String tag = entry.value;
-                          return _buildTag(tag, tagColors[tagIndex % tagColors.length], textColor[tagIndex % tagColors.length]); // Assigning colors cyclically
-                        }).toList(),
+                        children:
+                            meal["tags"].asMap().entries.map<Widget>((entry) {
+                              int tagIndex = entry.key;
+                              String tag = entry.value;
+                              return _buildTag(
+                                tag,
+                                tagColors[tagIndex % tagColors.length],
+                                textColor[tagIndex % tagColors.length],
+                              ); // Assigning colors cyclically
+                            }).toList(),
                       ),
                       SizedBox(height: 5),
-                      Text(meal["name"], style: TextStyle(fontSize: 16, fontWeight: FontWeight.w600)),
-                      Text(meal["details"], style: TextStyle(fontSize: 12, color: Color(0xFFB2B2B2))),
+                      Text(
+                        meal["name"],
+                        style: TextStyle(
+                          fontSize: 16,
+                          fontWeight: FontWeight.w600,
+                          fontFamily: 'Nunito',
+                        ),
+                      ),
+                      Text(
+                        meal["details"],
+                        style: TextStyle(
+                          fontSize: 12,
+                          color: Color(0xFFB2B2B2),
+                          fontFamily: 'Nunito',
+                        ),
+                      ),
                       SizedBox(height: 8),
                       Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          Text("£${meal["price"].toStringAsFixed(2)}", style: TextStyle(fontSize: 36, fontWeight: FontWeight.bold, color: Color(0xFF2B2D42))),
+                          Text(
+                            "£${meal["price"].toStringAsFixed(2)}",
+                            style: TextStyle(
+                              fontSize: 36,
+                              fontWeight: FontWeight.bold,
+                              color: Color(0xFF2B2D42),
+                              fontFamily: 'Nunito',
+                            ),
+                          ),
                           Row(
                             children: [
-                              _counterButton(PhosphorIcons.minus, () {
-                                if (mealCount > 0) {
-                                  setState(() {
-                                    mealCount--;
-                                  });
-                                }
-                              },
-                                buttonColor: mealCount != 0 ? Color(0xFFCF353F) : Color.fromRGBO(207, 53, 63, 0.5),
+                              _counterButton(
+                                PhosphorIcons.minus,
+                                () {
+                                  if (mealCount > 0) {
+                                    setState(() {
+                                      mealCount--;
+                                    });
+                                  }
+                                },
+                                buttonColor:
+                                    mealCount != 0
+                                        ? Color(0xFFCF353F)
+                                        : Color.fromRGBO(207, 53, 63, 0.5),
                               ),
                               Padding(
-                                padding: const EdgeInsets.symmetric(horizontal: 8),
+                                padding: const EdgeInsets.symmetric(
+                                  horizontal: 8,
+                                ),
                                 child: Text(
                                   "$mealCount",
-                                  style: const TextStyle(fontSize: 36, fontWeight: FontWeight.bold),
+                                  style: const TextStyle(
+                                    fontSize: 36,
+                                    fontWeight: FontWeight.bold,
+                                    fontFamily: 'Nunito',
+                                  ),
                                 ),
                               ),
-                              _counterButton(Icons.add, () {
-                                if (mealCount < 10) {
-                                  setState(() {
-                                    mealCount ++;
-                                  });
-                                }},
-                                buttonColor: mealCount != 10 ? Color(0xFFCF353F) : Color.fromRGBO(207, 53, 63, 0.5),
+                              _counterButton(
+                                Icons.add,
+                                () {
+                                  if (mealCount < 10) {
+                                    setState(() {
+                                      mealCount++;
+                                    });
+                                  }
+                                },
+                                buttonColor:
+                                    mealCount != 10
+                                        ? Color(0xFFCF353F)
+                                        : Color.fromRGBO(207, 53, 63, 0.5),
                               ),
                             ],
                           ),
                         ],
-                      )
+                      ),
                     ],
                   ),
                 ),
@@ -294,47 +392,67 @@ class _AddMealsPageState extends State<AddMealsPage> {
         color: color,
         borderRadius: BorderRadius.circular(100),
       ),
-      child: Text(label, style: TextStyle(color: textColor, fontSize: 10)),
+      child: Text(
+        label,
+        style: TextStyle(color: textColor, fontSize: 10, fontFamily: 'Nunito'),
+      ),
     );
   }
 
   Widget _buildBottomBar() {
     return Container(
       padding: EdgeInsets.symmetric(horizontal: 16, vertical: 14),
-      decoration: BoxDecoration(
-        color: Colors.white
-      ),
+      decoration: BoxDecoration(color: Colors.white),
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text("Total Meals ", style: TextStyle(fontSize: 12, color: Color(0xFF757575))),
-              Text("$mealCount/10", style: TextStyle(fontSize: 24, color: Color(0xFF2B2D42), fontWeight: FontWeight.w600)),
+              Text(
+                "Total Meals ",
+                style: TextStyle(
+                  fontSize: 12,
+                  color: Color(0xFF757575),
+                  fontFamily: 'Nunito',
+                ),
+              ),
+              Text(
+                "$mealCount/10",
+                style: TextStyle(
+                  fontSize: 24,
+                  color: Color(0xFF2B2D42),
+                  fontWeight: FontWeight.w600,
+                  fontFamily: 'Nunito',
+                ),
+              ),
             ],
           ),
           SizedBox(
             width: 200,
-            child: CustomButton(onTap: (){
-              Navigator.pushNamed(context, '/confirm');
-            }, text: "Continue"),
-          )
+            child: CustomButton(
+              onTap: () {
+                Navigator.pushNamed(context, '/confirm');
+              },
+              text: "Continue",
+            ),
+          ),
         ],
       ),
     );
   }
 
-  Widget _counterButton(IconData icon, VoidCallback onPressed, {Color buttonColor = const Color(0xFFCF353F)}) {
+  Widget _counterButton(
+    IconData icon,
+    VoidCallback onPressed, {
+    Color buttonColor = const Color(0xFFCF353F),
+  }) {
     return InkWell(
       onTap: onPressed,
       borderRadius: BorderRadius.circular(20),
       child: Container(
         padding: const EdgeInsets.all(8),
-        decoration: BoxDecoration(
-          color: buttonColor,
-          shape: BoxShape.circle,
-        ),
+        decoration: BoxDecoration(color: buttonColor, shape: BoxShape.circle),
         child: Icon(icon, color: Colors.white, size: 24),
       ),
     );
