@@ -2,19 +2,25 @@ import 'package:flutter/material.dart';
 import 'package:flutter_phosphor_icons/flutter_phosphor_icons.dart';
 
 class EditProfileScreen extends StatelessWidget {
+  const EditProfileScreen({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text(
           'Edit Profile',
-          style: TextStyle(fontSize: 18, fontWeight: FontWeight.w500),
+          style: TextStyle(
+            fontSize: 18,
+            fontWeight: FontWeight.w500,
+            fontFamily: 'Nunito',
+          ),
         ),
         centerTitle: true,
         backgroundColor: Colors.white,
         elevation: 0,
         leading: IconButton(
-          icon: Icon(PhosphorIcons.caret_left, size: 24, ),
+          icon: Icon(PhosphorIcons.caret_left, size: 24),
           onPressed: () => Navigator.pop(context),
         ),
       ),
@@ -30,7 +36,9 @@ class EditProfileScreen extends StatelessWidget {
                 children: [
                   CircleAvatar(
                     radius: 50,
-                    backgroundImage: AssetImage('assets/profile.jpg'), // Replace with network image if needed
+                    backgroundImage: AssetImage(
+                      'assets/profile.jpg',
+                    ), // Replace with network image if needed
                   ),
                   CircleAvatar(
                     radius: 14,
@@ -41,21 +49,31 @@ class EditProfileScreen extends StatelessWidget {
               ),
             ),
             SizedBox(height: 20),
-            Text("Name", style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold)),
+            Text(
+              "Name",
+              style: TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.bold,
+                fontFamily: 'Nunito',
+              ),
+            ),
             SizedBox(height: 5),
             Row(
               children: [
-                Expanded(
-                  child: _buildTextField(PhosphorIcons.user, "John"),
-                ),
+                Expanded(child: _buildTextField(PhosphorIcons.user, "John")),
                 SizedBox(width: 10),
-                Expanded(
-                  child: _buildTextField(PhosphorIcons.user, "Doe"),
-                ),
+                Expanded(child: _buildTextField(PhosphorIcons.user, "Doe")),
               ],
             ),
             SizedBox(height: 20),
-            Text("Phone Number", style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold)),
+            Text(
+              "Phone Number",
+              style: TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.bold,
+                fontFamily: 'Nunito',
+              ),
+            ),
             SizedBox(height: 5),
             _buildTextField(PhosphorIcons.phone, "+44 98 8393 9002"),
             SizedBox(height: 40),
@@ -90,11 +108,19 @@ class EditProfileScreen extends StatelessWidget {
         },
         style: ElevatedButton.styleFrom(
           backgroundColor: Colors.red,
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+          ),
           padding: EdgeInsets.symmetric(vertical: 16),
         ),
-        child: Text("Update Profile",
-            style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold, color: Colors.white)),
+        child: Text(
+          "Update Profile",
+          style: TextStyle(
+            fontSize: 16,
+            fontWeight: FontWeight.bold,
+            color: Colors.white,
+          ),
+        ),
       ),
     );
   }
