@@ -2,8 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_phosphor_icons/flutter_phosphor_icons.dart';
 
 class NotificationSettingsPage extends StatefulWidget {
+  const NotificationSettingsPage({super.key});
+
   @override
-  _NotificationSettingsPageState createState() => _NotificationSettingsPageState();
+  _NotificationSettingsPageState createState() =>
+      _NotificationSettingsPageState();
 }
 
 class _NotificationSettingsPageState extends State<NotificationSettingsPage> {
@@ -15,16 +18,20 @@ class _NotificationSettingsPageState extends State<NotificationSettingsPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar:  AppBar(
+      appBar: AppBar(
         title: Text(
           'Notification Settings',
-          style: TextStyle(fontSize: 18, fontWeight: FontWeight.w500),
+          style: TextStyle(
+            fontSize: 18,
+            fontWeight: FontWeight.w500,
+            fontFamily: 'Nunito',
+          ),
         ),
         centerTitle: true,
         backgroundColor: Colors.white,
         elevation: 0,
         leading: IconButton(
-          icon: Icon(PhosphorIcons.caret_left, size: 24, ),
+          icon: Icon(PhosphorIcons.caret_left, size: 24),
           onPressed: () => Navigator.pop(context),
         ),
       ),
@@ -34,28 +41,32 @@ class _NotificationSettingsPageState extends State<NotificationSettingsPage> {
           children: [
             _buildSwitchTile(
               title: "Delivery Updates",
-              subtitle: "Get real-time notifications about your upcoming meal deliveries.",
+              subtitle:
+                  "Get real-time notifications about your upcoming meal deliveries.",
               value: deliveryUpdates,
               onChanged: (val) => setState(() => deliveryUpdates = val),
             ),
-            SizedBox(height: 7,),
+            SizedBox(height: 7),
             _buildSwitchTile(
               title: "Subscription Renewal",
-              subtitle: "Stay informed about your subscription renewal dates and payment reminders.",
+              subtitle:
+                  "Stay informed about your subscription renewal dates and payment reminders.",
               value: subscriptionRenewal,
               onChanged: (val) => setState(() => subscriptionRenewal = val),
             ),
-            SizedBox(height: 7,),
+            SizedBox(height: 7),
             _buildSwitchTile(
               title: "Order Updates",
-              subtitle: "Receive alerts on order confirmations, status changes, and cancellations.",
+              subtitle:
+                  "Receive alerts on order confirmations, status changes, and cancellations.",
               value: orderUpdates,
               onChanged: (val) => setState(() => orderUpdates = val),
             ),
-            SizedBox(height: 7,),
+            SizedBox(height: 7),
             _buildSwitchTile(
               title: "App Updates",
-              subtitle: "Be the first informed about enhancements, new features, and critical updates.",
+              subtitle:
+                  "Be the first informed about enhancements, new features, and critical updates.",
               value: appUpdates,
               onChanged: (val) => setState(() => appUpdates = val),
             ),
@@ -65,15 +76,30 @@ class _NotificationSettingsPageState extends State<NotificationSettingsPage> {
     );
   }
 
-  Widget _buildSwitchTile({required String title, required String subtitle, required bool value, required Function(bool) onChanged}) {
+  Widget _buildSwitchTile({
+    required String title,
+    required String subtitle,
+    required bool value,
+    required Function(bool) onChanged,
+  }) {
     return Container(
       decoration: BoxDecoration(
         color: Colors.white,
-          borderRadius: BorderRadius.circular(10)
+        borderRadius: BorderRadius.circular(10),
       ),
       child: SwitchListTile(
-        title: Text(title, style: TextStyle(fontWeight: FontWeight.w500, fontSize: 16)),
-        subtitle: Text(subtitle, style: TextStyle(fontSize: 14),),
+        title: Text(
+          title,
+          style: TextStyle(
+            fontWeight: FontWeight.w500,
+            fontSize: 16,
+            fontFamily: 'Nunito',
+          ),
+        ),
+        subtitle: Text(
+          subtitle,
+          style: TextStyle(fontSize: 14, fontFamily: 'Nunito'),
+        ),
         value: value,
         onChanged: onChanged,
         activeTrackColor: Color(0xFFCF353F),
