@@ -1,80 +1,69 @@
-import 'package:app/custom_button.dart';
+import 'package:frontend_flutter/custom_button.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_phosphor_icons/flutter_phosphor_icons.dart';
 
 class SubscriptionDetailsPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar:  AppBar(
-        title: Text(
-          'Subscription Details',
-          style: TextStyle(fontSize: 18, fontWeight: FontWeight.w500),
-        ),
-        centerTitle: true,
-        backgroundColor: Colors.white,
-        elevation: 0,
-        leading: IconButton(
-          icon: Icon(PhosphorIcons.caret_left, size: 24, ),
-          onPressed: () => Navigator.pop(context),
-        ),
-      ),
-      body: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          _buildPlanSection(),
-          SizedBox(height: 20),
-          _buildInvoiceSection(),
-          Spacer(),
-          Container(
-            padding: EdgeInsets.all(16),
-            color: Colors.white,
-            child: CustomButton(
-              onTap: (){},
-              text: 'Renew',
-              backgroundColor: Color.fromRGBO(207, 53, 63, 0.5),
-              borderColor: Colors.transparent,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
+  const SubscriptionDetailsPage({super.key});
 
-  Widget _buildPlanSection() {
-    return  Padding(
-      padding: EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+  Widget _buildPlanSection(BuildContext context) {
+    final screenWidth = MediaQuery.of(context).size.width;
+    final horizontalPadding = screenWidth * 0.04;
+    final buttonWidth = (screenWidth - (horizontalPadding * 2) - 16) / 2;
+
+    return Padding(
+      padding: EdgeInsets.symmetric(
+        horizontal: horizontalPadding,
+        vertical: screenWidth * 0.025,
+      ),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           Text(
             'Your Plan - General',
-            style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+            style: TextStyle(
+              fontSize: 18,
+              fontWeight: FontWeight.bold,
+              fontFamily: 'Nunito',
+            ),
+            overflow: TextOverflow.ellipsis,
           ),
-          SizedBox(height: 10),
+          SizedBox(height: screenWidth * 0.025),
           _buildPlanDetail('Plan Starts', '22 January, 2025'),
           _buildPlanDetail('Plan Ends', '22 February, 2025'),
           _buildPlanDetail('Upcoming Delivery', 'Today at 12:30 PM'),
           _buildPlanDetail('Next Renewal', '21 February, 2025'),
-          SizedBox(height: 15),
+          SizedBox(height: screenWidth * 0.04),
           Container(
-            padding: EdgeInsets.all(16),
+            padding: EdgeInsets.all(horizontalPadding),
             decoration: BoxDecoration(
               color: Colors.white,
-              borderRadius: BorderRadius.circular(12)
+              borderRadius: BorderRadius.circular(12),
             ),
             child: Column(
               children: [
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                    Text(
-                      'Remaining Meals',
-                      style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600, color: Color(0xFF2B2D42)),
+                    Expanded(
+                      child: Text(
+                        'Remaining Meals',
+                        style: TextStyle(
+                          fontSize: 14,
+                          fontWeight: FontWeight.w600,
+                          color: Color(0xFF2B2D42),
+                          fontFamily: 'Nunito',
+                        ),
+                        overflow: TextOverflow.ellipsis,
+                      ),
                     ),
                     Text(
                       '08/10',
-                      style: TextStyle(fontSize: 14, fontWeight: FontWeight.bold),
+                      style: TextStyle(
+                        fontSize: 14,
+                        fontWeight: FontWeight.bold,
+                        fontFamily: 'Nunito',
+                      ),
                     ),
                   ],
                 ),
@@ -87,23 +76,23 @@ class SubscriptionDetailsPage extends StatelessWidget {
               ],
             ),
           ),
-          SizedBox(height: 15),
+          SizedBox(height: screenWidth * 0.04),
           Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
               CustomButton(
-                onTap: () {} ,
+                onTap: () {},
                 text: 'Pause',
                 backgroundColor: Color(0xFFFFCC3D),
-                width: 170,
+                width: buttonWidth,
                 textColor: Color(0xFF2B2D42),
                 borderColor: Colors.transparent,
               ),
               CustomButton(
-                onTap: () {} ,
+                onTap: () {},
                 text: 'Cancel',
                 backgroundColor: Color(0xFFFF4949),
-                width: 170,
+                width: buttonWidth,
                 borderColor: Colors.transparent,
               ),
             ],
@@ -119,18 +108,46 @@ class SubscriptionDetailsPage extends StatelessWidget {
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
-          Text(title, style: TextStyle(fontSize: 14, color: Color(0xFF2B2D42))),
-          Text(value, style: TextStyle(fontSize: 14, fontWeight: FontWeight.w500, color: Color(0xFF2B2D42))),
+          Expanded(
+            child: Text(
+              title,
+              style: TextStyle(
+                fontSize: 14,
+                color: Color(0xFF2B2D42),
+                fontFamily: 'Nunito',
+              ),
+              overflow: TextOverflow.ellipsis,
+            ),
+          ),
+          Expanded(
+            child: Text(
+              value,
+              style: TextStyle(
+                fontSize: 14,
+                fontWeight: FontWeight.w500,
+                color: Color(0xFF2B2D42),
+                fontFamily: 'Nunito',
+              ),
+              overflow: TextOverflow.ellipsis,
+              textAlign: TextAlign.end,
+            ),
+          ),
         ],
       ),
     );
   }
 
-  Widget _buildInvoiceSection() {
+  Widget _buildInvoiceSection(BuildContext context) {
+    final screenWidth = MediaQuery.of(context).size.width;
+    final horizontalPadding = screenWidth * 0.04;
+
     return Padding(
-      padding: EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+      padding: EdgeInsets.symmetric(
+        horizontal: horizontalPadding,
+        vertical: screenWidth * 0.025,
+      ),
       child: Container(
-        padding: EdgeInsets.all(16),
+        padding: EdgeInsets.all(horizontalPadding),
         decoration: BoxDecoration(
           color: Colors.white,
           borderRadius: BorderRadius.circular(10),
@@ -140,14 +157,82 @@ class SubscriptionDetailsPage extends StatelessWidget {
           children: [
             Text(
               'Invoice History',
-              style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+              style: TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.bold,
+                fontFamily: 'Nunito',
+              ),
+              overflow: TextOverflow.ellipsis,
             ),
-            Divider(thickness: 1, height: 5, color: Color(0xFFE0E0E0),),
-            SizedBox(height: 3,),
+            Divider(thickness: 1, height: 5, color: Color(0xFFE0E0E0)),
+            SizedBox(height: screenWidth * 0.01),
             _buildPlanDetail('Last Amount Paid', '£49.86'),
             _buildPlanDetail('Date', '22 January, 2025'),
             _buildPlanDetail('Payment Method', 'Apple Pay'),
           ],
+        ),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final screenWidth = MediaQuery.of(context).size.width;
+    final horizontalPadding = screenWidth * 0.04;
+
+    return AnnotatedRegion<SystemUiOverlayStyle>(
+      value: SystemUiOverlayStyle(
+        systemNavigationBarColor: Colors.white,
+        systemNavigationBarIconBrightness: Brightness.dark,
+      ),
+      child: Scaffold(
+        appBar: AppBar(
+          title: Text(
+            'Subscription Details',
+            style: TextStyle(
+              fontSize: 18,
+              fontWeight: FontWeight.w500,
+              fontFamily: 'Nunito',
+            ),
+            overflow: TextOverflow.ellipsis,
+          ),
+          centerTitle: true,
+          backgroundColor: Colors.white,
+          elevation: 0,
+          leading: IconButton(
+            icon: Icon(PhosphorIcons.caret_left, size: 24),
+            onPressed: () => Navigator.pop(context),
+          ),
+        ),
+        body: SafeArea(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Expanded(
+                child: SingleChildScrollView(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      _buildPlanSection(context),
+                      SizedBox(height: screenWidth * 0.05),
+                      _buildInvoiceSection(context),
+                    ],
+                  ),
+                ),
+              ),
+              Container(
+                width: double.infinity,
+                padding: EdgeInsets.all(horizontalPadding),
+                color: Colors.white,
+                child: CustomButton(
+                  onTap: () {},
+                  text: 'Renew',
+                  backgroundColor: Color.fromRGBO(207, 53, 63, 0.5),
+                  borderColor: Colors.transparent,
+                ),
+              ),
+            ],
+          ),
         ),
       ),
     );
